+++ lib/rails_stats/json_formatter.rb
@@ -18,7 +18,6 @@ module RailsStats
       @result << stat_hash("Total", @grand_total).merge(code_test_hash) if @grand_total
 
       @result << { "schema_stats" => schema_info }
-      @result << { "sti_stats" => print_sti_stat }
       @result << { "polymorphic_stats" => print_polymorphic_stats }
 
       @result
@@ -53,14 +52,6 @@ module RailsStats
           "m_over_c" => m_over_c.to_s,
           "loc_over_m" => loc_over_m.to_s
         }
-      end
-
-      def print_sti_stat
-        if calculator.sti
-          {
-            "sti_models_count" => calculator.sti,
-          }
-        end
       end
 
       def print_polymorphic_stats